requirements.txt
torch>=2.0
torchvision
timm
transformers
tfrecord[torch]
coca-pytorch
numpy
Pillow
# tensorflow>=2.12,<2.16
tensorflow>=2.16.1
scikit-learn